refactor: rename iOS WatermelonDB sources to NitromelonDB
Move native/ios/WatermelonDB to native/ios/NitromelonDB and rename the
umbrella header so CocoaPods imports are <NitromelonDB/NitromelonDB.h>.
Leave watermelondbProvideSyncJson as the C turbo-sync entry point.
